Rename modify traj protocol
