+++ gromacs/protocols/protocol_modify_trajectory.py
